WebService: Abfrage des Graphen auf einem Floor

+++ trunk/StudMap/StudMap.Core/Graph/Node.cs
@@ -4,8 +4,8 @@
     {
         public int Id { get; set; }
 
-        public int X { get; set; }
+        public decimal X { get; set; }
 
-        public int Y { get; set; }
+        public decimal Y { get; set; }
     }
 }
